move moonstone/Video to ui/Video
