+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -30,6 +30,7 @@ import {FloatingLayerContext} from '@enact/ui/FloatingLayer/FloatingLayerDecorat
 import Media from '@enact/ui/Media';
 import Slottable from '@enact/ui/Slottable';
 import Touchable from '@enact/ui/Touchable';
+import Video from '@enact/ui/Video';
 import equals from 'ramda/src/equals';
 import PropTypes from 'prop-types';
 import React from 'react';
@@ -48,7 +49,6 @@ import MediaSlider from './MediaSlider';
 import FeedbackContent from './FeedbackContent';
 import FeedbackTooltip from './FeedbackTooltip';
 import Times from './Times';
-import Video from './Video';
 
 import css from './VideoPlayer.module.less';
