+++ packages/ui/Video/VideoPlayer.module.less
@@ -1,0 +1,11 @@
+// Video.module.less
+//
+
+.video {
+	height: 100%;
+	width: 100%;
+}
+
+.preloadVideo {
+	display: none;
+}
