+++ packages/ui/Video/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "Video.js"
+}
